requirements.txt
Django>=5.0.3
requests>=2.31.0
beautifulsoup4>=4.12.0
opencv-python>=4.8.0
face-recognition>=1.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
lxml>=4.9.0
urllib3>=2.0.0
insightface>=0.7.3
faiss-cpu>=1.11.0
numpy>=2.2.0 